+++ app/server/db/getComments.sql
@@ -1,0 +1,1 @@
+SELECT * FROM comments WHERE post_id=$1;
