+++ app/src/components/createEvent/CreateEvent.css
@@ -64,13 +64,20 @@ justify-self:center;
 }
 .descriptionHeader{
     padding-top:32px;
-    padding-left:5;
+    left:5%;
     font-size:3rem;
     padding-bottom:32px;
+    position:relative;
+    z-index:5;
+   
 }
 
 .eventDescription{
-    padding-left:5%;
+   
+    position:relative;
+    left:5%;
+    z-index:5;
+   
   
 }
 
@@ -216,25 +223,29 @@ react-datepicker__time-list{
     width:100%;
 }
 .catalogue{
-    
+    position:relative;
+    left:5%;
+    z-index:5;
+   
 }
 .catalogueImageContainer{
-margin-left:5%;
+
 width:10vw;
 height:100px;
 border: 1px dashed #C5C5C5;
 }
 .catalogueItemName{
   font-size:1rem;
-  padding-left:5%;
+  
 }
 .catalogueItemAuctionId{
     font-size:1rem;
     color:red;
-    padding-left:5%;
+   
 }
 .catalogueNextPreviousContainer{
-
+  
+    position:relative;
     width:20vw;
     display:grid;
     grid-template-columns: 20% 60% 20%;
@@ -246,6 +257,7 @@ border: 1px dashed #C5C5C5;
     grid-row-start:1;
     grid-row-end:1;
     align-self:center;
+    justify-self:end;
     font-size:3rem;
 }
 .catalogueNext{
@@ -257,6 +269,8 @@ border: 1px dashed #C5C5C5;
     font-size:3rem;
 }
 .ImageContainer{
+    position:relative;
+    z-index:4;
    grid-column-start:2;
    grid-column-end:3;
    grid-row:1;
@@ -272,5 +286,5 @@ border: 1px dashed #C5C5C5;
 }
 .createEventsPage{
     position:relative;
-    z-index:4;
+   
 }
