+++ app/src/styles/BlogDetails.css
